AUT-2931: Move email fraud check logic
Email fraud check API call logic was previously in the enter email controller.
This commit moves it to the check your email controller where it should be.
