+++ src/components/enter-email/enter-email-controller.ts
@@ -19,7 +19,6 @@ import { checkReauthUsersService } from "../check-reauth-users/check-reauth-user
 import {
   getEmailEnteredWrongBlockDurationInMinutes,
   support2hrLockout,
-  supportCheckEmailFraud,
   supportReauthentication,
 } from "../../config";
 import {
@@ -32,9 +31,6 @@ import {
   timestampNMinutesFromNow,
   timestampNSecondsFromNow,
 } from "../../utils/lock-helper";
-import { CheckEmailFraudBlockInterface } from "../check-email-fraud-block/types";
-import { checkEmailFraudBlockService } from "../check-email-fraud-block/check-email-fraud-block-service";
-import { logger } from "../../utils/logger";
 
 export const RE_ENTER_EMAIL_TEMPLATE =
   "enter-email/index-re-enter-email-account.njk";
@@ -63,8 +59,7 @@ export function enterEmailCreateGet(req: Request, res: Response): void {
 
 export function enterEmailPost(
   service: EnterEmailServiceInterface = enterEmailService(),
-  checkReauthService: CheckReauthServiceInterface = checkReauthUsersService(),
-  checkEmailFraudService: CheckEmailFraudBlockInterface = checkEmailFraudBlockService()
+  checkReauthService: CheckReauthServiceInterface = checkReauthUsersService()
 ): ExpressRouteFunc {
   return async function (req: Request, res: Response) {
     const email = req.body.email;
@@ -131,17 +126,6 @@ export function enterEmailPost(
       result.data.lockoutInformation.length > 0
     )
       setUpAuthAppLocks(req, result.data.lockoutInformation);
-    if (supportCheckEmailFraud()) {
-      const checkEmailFraudResponse =
-        await checkEmailFraudService.checkEmailFraudBlock(
-          email,
-          sessionId,
-          clientSessionId,
-          persistentSessionId,
-          req
-        );
-      logger.info(`checkEmailFraudResponse: ${checkEmailFraudResponse.data}`);
-    }
     req.session.user.enterEmailMfaType = result.data.mfaMethodType;
     req.session.user.redactedPhoneNumber = result.data.phoneNumberLastThree;
     const nextState = result.data.doesUserExist
