+++ src/components/enter-email/tests/enter-email-integration.test.ts
@@ -21,7 +21,6 @@ describe("Integration::enter email", () => {
   let baseApi: string;
 
   before(async () => {
-    process.env.SUPPORT_CHECK_EMAIL_FRAUD = "1";
     decache("../../../app");
     decache("../../../middleware/session-middleware");
     const sessionMiddleware = require("../../../middleware/session-middleware");
@@ -168,13 +167,6 @@ describe("Integration::enter email", () => {
         email: "test@test.com",
         doesUserExist: true,
       });
-    nock(baseApi)
-      .post(API_ENDPOINTS.CHECK_EMAIL_FRAUD_BLOCK)
-      .once()
-      .reply(HTTP_STATUS_CODES.OK, {
-        email: "test@test.com",
-        isBlockedStatus: "Pending",
-      });
 
     request(app)
       .post(PATH_NAMES.ENTER_EMAIL_SIGN_IN)
@@ -193,13 +185,6 @@ describe("Integration::enter email", () => {
       email: "test@test.com",
       doesUserExist: false,
     });
-    nock(baseApi)
-      .post(API_ENDPOINTS.CHECK_EMAIL_FRAUD_BLOCK)
-      .once()
-      .reply(HTTP_STATUS_CODES.OK, {
-        email: "test@test.com",
-        isBlockedStatus: "Pending",
-      });
 
     request(app)
       .post(PATH_NAMES.ENTER_EMAIL_SIGN_IN)
@@ -249,14 +234,6 @@ describe("Integration::enter email", () => {
       .reply(HTTP_STATUS_CODES.OK, {
         email: "test@test.com",
         doesUserExist: true,
-      });
-
-    nock(baseApi)
-      .post(API_ENDPOINTS.CHECK_EMAIL_FRAUD_BLOCK)
-      .once()
-      .reply(HTTP_STATUS_CODES.OK, {
-        email: "test@test.com",
-        isBlockedStatus: "Pending",
       });
 
     request(app)
